utils/forms refactoring, pt 1
