+++ calvincTools/deprecated code/cQdbSimpleRecordFormWidgets.py
