+++ calvincTools/deprecated code/cQdbSimpleRecordSubFormWidgets.old-py-code
