+++ calvincTools/utils/forms/definitions/cQFormBtnDef.py
