+++ calvincTools/utils/forms/definitions/cQFormFieldDef.py
@@ -5,7 +5,7 @@ from enum import Enum
 from PySide6.QtCore import Qt
 from PySide6.QtWidgets import QWidget
 
-from .cQFormWidgets import cQFmConstants
+from .cQFmConstants import cQFmConstants
 
 @dataclass(frozen=True)
 class cQFormFieldDef:
